+++ src/math/min_max.c
@@ -6,11 +6,11 @@
 /*   By: gastesan <gastesan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/22 02:28:23 by gastesan          #+#    #+#             */
-/*   Updated: 2026/01/15 14:24:31 by gastesan         ###   ########.fr       */
+/*   Updated: 2026/01/15 14:36:19 by gastesan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	min(int a, int b)
+long	min(long a, long b)
 {
 	if (a < b)
 		return (a);
@@ -18,7 +18,7 @@ int	min(int a, int b)
 		return (b);
 }
 
-int	max(int a, int b)
+long	max(long a, long b)
 {
 	if (a > b)
 		return (a);
